Fix C4244 (conversion from double to float)
I imagine the use of `std::stod` was in error, though I can't be entirely certain.

Running msbuild now produces 3 warnings, down from 56.

+++ thprac/src/thprac/thprac_launcher_cfg.cpp
@@ -290,7 +290,7 @@ void LauncherSettingSet(const char* name, std::string& valueIn)
             if (jsonVec[i].IsNumber()) {                                        \
                 val = jsonVec[i].GetFloat();                                    \
             } else if (jsonVec[i].IsString()) {                                 \
-                val = std::stod(jsonVec[i].GetString());                        \
+                val = std::stof(jsonVec[i].GetString());                        \
             } else {                                                            \
                 throw std::runtime_error(#col ": wrong format");                \
             }                                                                   \
